requirements.txt
python-telegram-bot==21.6
gallery-dl==1.29.6
